backend/requirements.txt
fastapi
pydantic
chess
stockfish
python-dotenv
uvicorn